Moved tests out of source folder Replaced be.ok as it's not recommended: https://www.chaijs.com/api/bdd/#method_ok Fixed order in _setAsBox to be consistent with wrapping algorithm (so vertices are not reordered for Box fixture) Commented out some debug output
